--- backend/src/main/kotlin/com/jetbrains/life_science/article/content/index/IndexCoordinatesProvider.kt
@@ -1,8 +1,0 @@
-package com.jetbrains.life_science.article.content.index
-
-import org.springframework.data.elasticsearch.core.mapping.IndexCoordinates
-
-interface IndexCoordinatesProvider {
-
-    val indexCoordinates: IndexCoordinates
-}
