+++ backend/src/test/kotlin/com/jetbrains/life_science/util/populator/ElasticPopulator.kt
@@ -23,7 +23,7 @@ class ElasticPopulator(
 
     fun prepareData() {
         populators.forEach { it.prepareData() }
-        runBlocking { delay(400) }
+        runBlocking { delay(100) }
     }
 
     fun createIndexes() {
